My sorting algorithm/function :)
I'm relativly certain that there is an improvement that I could do to optimize this function.
- templates
- pointers
- treating arrays as pointers (because they are)
- I would consider this a somewhat advanced use of pointers/arrays, at least in the scope of this repo.
- operators
- loops
- if-statements
- functions

- edit: this now supports 32-bit windows

+++ sort_array.cpp
@@ -32,7 +32,7 @@ int main(){
   std::cin >>length;
 
   // allocate some memory to hold our array
-  long double userArray[length];
+  double userArray[length];
 
   std::cout <<"Enter " <<length <<" numbers:\n";
